requirements.txt
supervisor
eventlet
beautifulsoup4==4.12.3
# comtypes==1.4.9
Django==5.1.2
django_celery_results==2.5.1
duckduckgo_search==7.3.0
PyMuPDF==1.26.0
fuzzywuzzy==0.18.0
langchain==0.3.12
langchain_community==0.3.12
langchain_core==0.3.25
langchain_huggingface==0.1.2
langchain_openai==0.2.12
langchain_text_splitters==0.3.3
language_tool_python==2.8.2
numpy==1.26.4
opencv_contrib_python==4.10.0.84
opencv_python==4.11.0.86
opencv_python_headless==4.10.0.84
openpyxl==3.1.5
pandas==2.2.3
patool==3.1.0
pdfplumber==0.11.4
pillow==10.4.0
psycopg2-binary==2.9.10
pyproj==3.7.1
pytesseract==0.3.13
python-docx==1.1.2
python-Levenshtein==0.26.1
# pywin32==308
redis==5.2.1
requests==2.32.3
setuptools==75.6.0
shapely==2.0.6
simplekml==1.3.6
scikit-image==0.24.0
tensorflow==2.18.0
tensorflow_cpu>=2.18.0,<2.19.0
transliterate==1.10.2
typing_extensions==4.12.2

pydantic==2.9.2
pydantic-settings==2.7.0
pydantic_core==2.23.4
aiohttp==3.10.10
aiohttp-cors==0.7.0
python-dateutil==2.8.2
click #==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django-Accounts==0.1
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-nested-inline==0.4.6
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
pytz==2024.2
Markdown==3.7
markdown-it-py==3.0.0
channels==4.2.0
channels_redis==4.2.1
celery==5.4.0
celery-progress==0.4
tk==0.1.0